refactor: 카테고리 페이지 코드 리팩토링
